requirements.txt
torch==2.5.0
torchvision==0.20.0
streamlit==1.50.0
Pillow==11.3.0
